SERIAL_PORT/pyproject.toml
[tool.poetry]
name = "serial-port"
version = "0.1.0"
description = ""
authors = ["SzymonGogulski <szymongogulskiedu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.1.4"
matplotlib = "^3.8.2"
pyserial = "^3.5"
control = "^0.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
